requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
pandas==2.1.0
python-dotenv==1.0.0
requests==2.31.0

# Database
sqlalchemy==2.0.0
psycopg2-binary==2.9.7
alembic==1.12.0
